Added error checking function to remove outliers

+++ src/main/java/io/github/droshux/rigidbodies/Point.java
@@ -7,17 +7,18 @@ public class Point {
     public double y;
 
     public Point(double X, double Y) {
-        this.x = X; this.y = Y;
+        this.x = X;
+        this.y = Y;
     }
 
-    //Gets the distance between two points
-    public float DistanceTo(Point other) {
-        double deltaX = Math.abs(this.x-other.x);
-        double deltaY = Math.abs(this.y-other.y);
-        return (float)Math.sqrt(Math.pow(deltaX, 2) + Math.pow(deltaY, 2));
+    // Gets the distance between two points
+    public double DistanceTo(Point other) {
+        double deltaX = Math.abs(this.x - other.x);
+        double deltaY = Math.abs(this.y - other.y);
+        return Math.sqrt(Math.pow(deltaX, 2) + Math.pow(deltaY, 2));
     }
 
-    //To string override :)
+    // To string override :)
 
     @Override
     public String toString() {
@@ -25,27 +26,32 @@ public class Point {
     }
 
     public void matrixTransformSelf(double ix, double jx,
-                                double iy, double jy) {
+            double iy, double jy) {
         double xOUT = (ix * x) + (jx * y);
-        double yOUT = (iy * x) + (jy  * y);
+        double yOUT = (iy * x) + (jy * y);
 
-        x = xOUT; y=yOUT;
+        x = xOUT;
+        y = yOUT;
     }
+
     @SuppressWarnings("unused")
     public void matrixTransformSelf(Utils.Matrix M) {
         double xOUT = (M.i.x * x) + (M.j.x * y);
         double yOUT = (M.i.y * x) + (M.j.y * y);
 
-        x=xOUT;y=yOUT;
+        x = xOUT;
+        y = yOUT;
     }
+
     @SuppressWarnings("unused")
     public Point matrixTransform(double ix, double jx,
-                                double iy, double jy) {
+            double iy, double jy) {
         double xOUT = (ix * x) + (jx * y);
-        double yOUT = (iy * x) + (jy  * y);
+        double yOUT = (iy * x) + (jy * y);
 
         return new Point(xOUT, yOUT);
     }
+
     @SuppressWarnings("unused")
     public Point matrixTransform(Utils.Matrix M) {
         double xOUT = (M.i.x * x) + (M.j.x * y);
@@ -56,8 +62,10 @@ public class Point {
 
     @Override
     public boolean equals(Object o) {
-        if (this == o) return true;
-        if (!(o instanceof Point point)) return false;
+        if (this == o)
+            return true;
+        if (!(o instanceof Point point))
+            return false;
         return Double.compare(point.x, x) == 0 && Double.compare(point.y, y) == 0;
     }
